+++ src/sll/include/sll/memory.h
@@ -47,6 +47,18 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT void* sll_allocate(sll_size_t sz);
 
 /**
  * \flags check_output func
+ * \name sll_allocate_fail
+ * \group memory
+ * \desc Docs!
+ * \arg sll_size_t sz
+ * \ret void*
+ */
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT void* sll_allocate_fail(sll_size_t sz);
+
+
+
+/**
+ * \flags check_output func
  * \name sll_allocate_stack
  * \group memory
  * \desc Docs!
@@ -54,6 +66,18 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT void* sll_allocate(sll_size_t sz);
  * \ret void*
  */
 __SLL_EXTERNAL __SLL_CHECK_OUTPUT void* sll_allocate_stack(sll_size_t sz);
+
+
+
+/**
+ * \flags check_output func
+ * \name sll_allocate_stack_fail
+ * \group memory
+ * \desc Docs!
+ * \arg sll_size_t sz
+ * \ret void*
+ */
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT void* sll_allocate_stack_fail(sll_size_t sz);
 
 
 
@@ -96,6 +120,19 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT void* sll_reallocate(void* p,sll_size_t sz);
 
 /**
  * \flags check_output func
+ * \name sll_reallocate_fail
+ * \group memory
+ * \desc Docs!
+ * \arg void* p
+ * \arg sll_size_t sz
+ * \ret void*
+ */
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT void* sll_reallocate_fail(void* p,sll_size_t sz);
+
+
+
+/**
+ * \flags check_output func
  * \name sll_zero_allocate
  * \group memory
  * \desc Docs!
@@ -108,6 +145,18 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT void* sll_zero_allocate(sll_size_t sz);
 
 /**
  * \flags check_output func
+ * \name sll_zero_allocate_fail
+ * \group memory
+ * \desc Docs!
+ * \arg sll_size_t sz
+ * \ret void*
+ */
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT void* sll_zero_allocate_fail(sll_size_t sz);
+
+
+
+/**
+ * \flags check_output func
  * \name sll_zero_allocate_stack
  * \group memory
  * \desc Docs!
@@ -115,6 +164,18 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT void* sll_zero_allocate(sll_size_t sz);
  * \ret void*
  */
 __SLL_EXTERNAL __SLL_CHECK_OUTPUT void* sll_zero_allocate_stack(sll_size_t sz);
+
+
+
+/**
+ * \flags check_output func
+ * \name sll_zero_allocate_stack_fail
+ * \group memory
+ * \desc Docs!
+ * \arg sll_size_t sz
+ * \ret void*
+ */
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT void* sll_zero_allocate_stack_fail(sll_size_t sz);